update bip39 and add support for more languages
