+++ cardano/src/bip/bip39_english.txt
