+++ cardano/src/bip/mod.rs
@@ -1,0 +1,8 @@
+//! BitCoin improvements proposal implementations
+//!
+//! This intends to provide convenient tooling for user of the cardano
+//! library for managing wallet, private keys and addresses.
+//!
+
+pub mod bip39;
+pub mod bip44;
