+++ cardano/src/paperwallet.rs
@@ -1,7 +1,6 @@
-extern crate cryptoxide;
-use self::cryptoxide::sha2::Sha512;
-use self::cryptoxide::hmac::Hmac;
-use self::cryptoxide::pbkdf2::{pbkdf2};
+use cryptoxide::sha2::Sha512;
+use cryptoxide::hmac::Hmac;
+use cryptoxide::pbkdf2::{pbkdf2};
 
 const ITERS : u32 = 10000;
 pub const IV_SIZE: usize = 8;
